src/app/requirements.txt
streamlit>=1.32.0
databricks-sdk>=0.81.0
sqlalchemy>=2.0
psycopg[binary]>=3.1
